requirements.txt
# Temporary requirements without browser-use for testing
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
supabase>=2.15.0
aiofiles>=24.1.0
aiohttp>=3.12.0
browser-use>=0.2.4
playwright>=1.40.0
fastmcp>=2.3.4
typer>=0.15.0
gotrue>=1.0.0
PyJWT>=2.8.0
python-dotenv>=1.0.0 